+++ com.unity.render-pipelines.universal/Runtime/UniversalRenderPipelineGlobalSettings.cs
@@ -4,6 +4,27 @@ using UnityEngine.Serialization;
 
 namespace UnityEngine.Rendering.Universal
 {
+    public enum ShaderVariantLogLevel
+    {
+        Disabled,
+        [InspectorName("Only URP Shaders")]
+        OnlyUniversalRPShaders,
+        [InspectorName("All Shaders")]
+        AllShaders
+    }
+    /// <summary>
+    /// Defines if Unity discards or stores the render targets of the DrawObjects Passes. Selecting the Store option significantly increases the memory bandwidth on mobile and tile-based GPUs.
+    /// </summary>
+    public enum StoreActionsOptimization
+    {
+        /// <summary>Unity uses the Discard option by default, and falls back to the Store option if it detects any injected Passes.</summary>
+        Auto,
+        /// <summary>Unity discards the render targets of render Passes that are not reused later (lower memory bandwidth).</summary>
+        Discard,
+        /// <summary>Unity stores all render targets of each Pass (higher memory bandwidth).</summary>
+        Store
+    }
+
     /// <summary>
     /// Universal Render Pipeline's Global Settings.
     /// Global settings are unique per Render Pipeline type. In URP, Global Settings contain:
@@ -172,6 +193,101 @@ namespace UnityEngine.Rendering.Universal
         {
             UpdateRenderingLayerNames();
         }
+
+        internal enum SpriteDefaultMaterialType
+        {
+            Lit,
+            Unlit,
+            Custom
+        }
+        [SerializeField] SpriteDefaultMaterialType m_DefaultSpriteMaterialType = SpriteDefaultMaterialType.Lit;
+        [SerializeField, Reload("Runtime/Materials/Sprite-Lit-Default.mat")] Material m_DefaultSpriteCustomMaterial = null;
+        public SpriteDefaultMaterialType defaultSpriteMaterialType { get { return m_DefaultSpriteMaterialType; } }
+        public Material defaultSpriteCustomMaterial { get { return m_DefaultSpriteCustomMaterial; } }
+
+        // General settings
+        [SerializeField] bool m_SupportsTerrainHoles = true;
+
+        public bool supportsTerrainHoles
+        {
+            get { return m_SupportsTerrainHoles; }
+        }
+
+
+        [SerializeField] [Obsolete] PipelineDebugLevel m_DebugLevel;
+        [SerializeField] ShaderVariantLogLevel m_ShaderVariantLogLevel = ShaderVariantLogLevel.Disabled;
+        public ShaderVariantLogLevel shaderVariantLogLevel
+        {
+            get { return m_ShaderVariantLogLevel; }
+            set { m_ShaderVariantLogLevel = value; }
+        }
+
+        [SerializeField] StoreActionsOptimization m_StoreActionsOptimization = StoreActionsOptimization.Auto;
+        /// <summary>
+        /// Returns the active store action optimization value.
+        /// </summary>
+        /// <returns>Returns the active store action optimization value.</returns>
+        public StoreActionsOptimization storeActionsOptimization
+        {
+            get { return m_StoreActionsOptimization; }
+            set { m_StoreActionsOptimization = value; }
+        }
+
+        // Post-processing settings
+        // Note: A lut size of 16^3 is barely usable with the HDR grading mode. 32 should be the
+        // minimum, the lut being encoded in log. Lower sizes would work better with an additional
+        // 1D shaper lut but for now we'll keep it simple.
+        public const int k_MinLutSize = 16;
+        public const int k_MaxLutSize = 65;
+        [SerializeField] ColorGradingMode m_ColorGradingMode = ColorGradingMode.LowDynamicRange;
+        [SerializeField] int m_ColorGradingLutSize = 32;
+        [SerializeField] bool m_UseFastSRGBLinearConversion = false;
+        [SerializeField] VolumeFrameworkUpdateMode m_VolumeFrameworkUpdateMode = VolumeFrameworkUpdateMode.EveryFrame;
+
+        public PostProcessData postProcessData = null;
+        public ColorGradingMode colorGradingMode
+        {
+            get { return m_ColorGradingMode; }
+            set { m_ColorGradingMode = value; }
+        }
+
+        public int colorGradingLutSize
+        {
+            get { return m_ColorGradingLutSize; }
+            set { m_ColorGradingLutSize = Mathf.Clamp(value, k_MinLutSize, k_MaxLutSize); }
+        }
+
+        /// <summary>
+        /// Returns true if fast approximation functions are used when converting between the sRGB and Linear color spaces, false otherwise.
+        /// </summary>
+        public bool useFastSRGBLinearConversion
+        {
+            get { return m_UseFastSRGBLinearConversion; }
+        }
+        /// <summary>
+        /// Returns the selected update mode for volumes.
+        /// </summary>
+        public VolumeFrameworkUpdateMode volumeFrameworkUpdateMode => m_VolumeFrameworkUpdateMode;
+
+        [SerializeField] bool m_UseNativeRenderPass = false;
+        public bool useNativeRenderPass
+        {
+            get => m_UseNativeRenderPass;
+            set
+            {
+                SetDirty();
+                m_UseNativeRenderPass = value;
+            }
+        }
+        [SerializeField]
+        TransparencySortMode m_TransparencySortMode = TransparencySortMode.Default;
+
+        [SerializeField]
+        Vector3 m_TransparencySortAxis = Vector3.up;
+
+        internal TransparencySortMode transparencySortMode => m_TransparencySortMode;
+        internal Vector3 transparencySortAxis => m_TransparencySortAxis;
+
 
         [System.NonSerialized]
         string[] m_RenderingLayerNames;
